tests/wpt/web-platform-tests/notifications/constructor-basic.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notification constructor (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    if (Notification.permission != "granted") {
        this.force_timeout()
        this.set_status(this.NOTRUN, "You must allow notifications for this"
            + " origin before running this test.")
    }
    var notification = new Notification("New Email Received")
    assert_true(notification instanceof Notification)
    notification.onshow = function() {
        notification.close()
    }
}, "Called the notification constructor with one argument.")
</script>
